+++ changelog.md
@@ -4,6 +4,7 @@
 
 ## Added
 
+* `[cellular]` Added call request type recognition.
 * `[calendar]` Added a new field for date of the event when the user adds/edits event.
 * `[file indexer db]` Added messages for File Indexer db.
 * `[settings]` Added Torch window (front-end only).